Move NodeEnums to Subscriber
