+++ BrewMesModulesParent/subscriber/pom.xml
@@ -27,6 +27,11 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.milo</groupId>
+			<artifactId>stack-core</artifactId>
+			<version>0.5.2</version>
+		</dependency>
 	</dependencies>
