+++ BrewMesModulesParent/subscriber/src/main/java/com/brewmes/subscriber/util/Constants.java
@@ -1,4 +1,4 @@
-package com.brewmes.common.util.node_enums;
+package com.brewmes.subscriber.util;
 
 class Constants {
     public static final int NAMESPACE_INDEX_VALUE = 6;
