+++ BrewMesModulesParent/subscriber/src/test/java/com/brewmes/subscriber/util/AdminNodesTest.java
@@ -1,4 +1,4 @@
-package com.brewmes.common.util.node_enums;
+package com.brewmes.subscriber.util;
 
 import org.junit.jupiter.api.Test;
 
@@ -8,11 +8,9 @@ class AdminNodesTest {
 
     @Test
     void values() {
-        assertNotNull(AdminNodes.values());
     }
 
     @Test
     void valueOf() {
-        assertEquals(AdminNodes.BATCH_PRODUCT_ID, AdminNodes.valueOf("BATCH_PRODUCT_ID"));
     }
 }
