+++ BrewMesModulesParent/subscriber/src/test/java/com/brewmes/subscriber/util/MachineNodesTest.java
@@ -1,4 +1,4 @@
-package com.brewmes.common.util.node_enums;
+package com.brewmes.subscriber.util;
 
 import org.junit.jupiter.api.Test;
